refactored the feedback system to be flatter structured
