+++ server/feedback/src/post_feedback.rs
@@ -1,34 +1,10 @@
-mod github;
-mod tokens;
-use crate::core::tokens::Claims;
+use crate::core::AppStateFeedback;
+use crate::{github, tokens};
 use actix_web::web::{Data, Json};
-use actix_web::{post, HttpResponse};
-use jsonwebtoken::{encode, EncodingKey, Header};
-use log::error;
+use actix_web::HttpResponse;
+
+use actix_web::post;
 use serde::Deserialize;
-
-use tokio::sync::Mutex;
-
-pub struct AppStateFeedback {
-    feedback_keys: crate::FeedbackKeys,
-    token_record: Mutex<Vec<TokenRecord>>,
-}
-impl AppStateFeedback {
-    pub fn from(feedback_keys: crate::FeedbackKeys) -> AppStateFeedback {
-        AppStateFeedback {
-            feedback_keys,
-            token_record: Mutex::new(Vec::new()),
-        }
-    }
-    pub fn able_to_process_feedback(&self) -> bool {
-        self.feedback_keys.github_token.is_some() && self.feedback_keys.jwt_key.is_some()
-    }
-}
-
-pub struct TokenRecord {
-    kid: u64,
-    next_reset: usize,
-}
 
 #[derive(Deserialize)]
 pub struct FeedbackPostData {
@@ -38,31 +14,6 @@ pub struct FeedbackPostData {
     body: String,
     privacy_checked: bool,
     delete_issue_requested: bool,
-}
-
-pub async fn get_token(state: Data<AppStateFeedback>) -> HttpResponse {
-    if !state.able_to_process_feedback() {
-        return HttpResponse::ServiceUnavailable()
-            .content_type("text/plain")
-            .body("Feedback is currently not configured on this server.");
-    }
-
-    let secret = state.feedback_keys.jwt_key.clone().unwrap(); // we checked available
-    let token = encode(
-        &Header::default(),
-        &Claims::new(),
-        &EncodingKey::from_secret(secret.as_bytes()),
-    );
-
-    match token {
-        Ok(token) => HttpResponse::Created().json(token),
-        Err(e) => {
-            error!("Failed to generate token: {e:?}");
-            HttpResponse::InternalServerError()
-                .content_type("text/plain")
-                .body("Failed to generate token.")
-        }
-    }
 }
 
 #[post("/api/feedback/feedback")]
@@ -90,10 +41,9 @@ pub async fn send_feedback(
         .unwrap()
         .trim()
         .to_string();
-    github::post_feedback(
+    github::open_issue(
         github_token,
-        title_category,
-        &req_data.subject,
+        &format!("[{title_category}] {subject}", subject = req_data.subject),
         &req_data.body,
         labels,
     )
